Fix for GC test mode assert not taking local storage flag into account
